Changed: Extract resizing methods to class in `ResizeBuffer`
Extract resizing methods to class. `UpdateOldBuffer` and `FastResize` class satisfy Single responsibility principle.
