+++ terminal-emulator/src/main/java/com/termux/terminal/ResizeBuffer.java
@@ -1,0 +1,5 @@
+package com.termux.terminal;
+
+public interface ResizeBuffer {
+    void resize(int newColumns, int newRows, int newTotalRows, int[] cursor, long currentStyle, boolean isAltScreen) ;
+}
