+++ terminal-emulator/src/main/java/com/termux/terminal/TerminalBuffer.java
@@ -16,9 +16,9 @@ public final class TerminalBuffer {
     /** The number of rows and columns visible on the screen. */
     int mScreenRows, mColumns;
     /** The number of rows kept in history. */
-    private int mActiveTranscriptRows = 0;
+    int mActiveTranscriptRows = 0;
     /** The index in the circular buffer where the visible screen starts. */
-    private int mScreenFirstRow = 0;
+    int mScreenFirstRow = 0;
 
     /**
      * Create a transcript screen.
@@ -125,199 +125,14 @@ public final class TerminalBuffer {
         // newRows > mTotalRows should not normally happen since mTotalRows is TRANSCRIPT_ROWS (10000):
         if (newColumns == mColumns && newRows <= mTotalRows) {
             // Fast resize where just the rows changed.
-            fastResize(newRows, newTotalRows, cursor, currentStyle, isAltScreen);
+            new FastResize(this).resize(newColumns, newRows, newTotalRows, cursor, currentStyle, isAltScreen);
         } else {
             // Copy away old state and update new:
-            updateOldState(newColumns, newRows, newTotalRows, cursor, currentStyle);
+            new UpdateOldBuffer(this).resize(newColumns, newRows, newTotalRows, cursor, currentStyle, isAltScreen);
         }
 
         // Handle cursor scrolling off screen:
         if (cursor[0] < 0 || cursor[1] < 0) cursor[0] = cursor[1] = 0;
-    }
-
-    private void fastResize(int newRows, int newTotalRows, int[] cursor, long currentStyle, boolean isAltScreen) {
-        int shiftDownOfTopRow = calcShiftDownOfTopRow(newRows, cursor, currentStyle);
-        mScreenFirstRow += shiftDownOfTopRow;
-        mScreenFirstRow = (mScreenFirstRow < 0) ? (mScreenFirstRow + mTotalRows) : (mScreenFirstRow % mTotalRows);
-        mTotalRows = newTotalRows;
-        mActiveTranscriptRows = isAltScreen ? 0 : Math.max(0, mActiveTranscriptRows + shiftDownOfTopRow);
-        cursor[1] -= shiftDownOfTopRow;
-        mScreenRows = newRows;
-    }
-
-    private int calcShiftDownOfTopRow(int newRows, int[] cursor, long currentStyle) {
-        int shiftDownOfTopRow = mScreenRows - newRows;
-        final boolean isShrinking = shiftDownOfTopRow > 0 && shiftDownOfTopRow < mScreenRows;
-        final boolean isExpanding = shiftDownOfTopRow < 0;
-        if (isShrinking) {
-            // Shrinking. Check if we can skip blank rows at bottom below cursor.
-            shiftDownOfTopRow = shrinkingRows(cursor, shiftDownOfTopRow);
-        } else if (isExpanding) {
-            // Negative shift down = expanding. Only move screen up if there is transcript to show:
-            shiftDownOfTopRow = expandingRows(currentStyle, shiftDownOfTopRow);
-        }
-        return shiftDownOfTopRow;
-    }
-
-    private int shrinkingRows(int[] cursor, int shiftDownOfTopRow) {
-        for (int i = mScreenRows - 1; i > 0; i--) {
-            if (cursor[1] >= i) break;
-            int r = externalToInternalRow(i);
-            final boolean isLineEmpty = mLines[r] == null || mLines[r].isBlank();
-            if (isLineEmpty) {
-                final boolean isShrinkingEnd = --shiftDownOfTopRow == 0;
-                if (isShrinkingEnd) break;
-            }
-        }
-        return shiftDownOfTopRow;
-    }
-
-    private int expandingRows(long currentStyle, int shiftDownOfTopRow) {
-        int actualShift = Math.max(shiftDownOfTopRow, -mActiveTranscriptRows);
-        if (shiftDownOfTopRow != actualShift) {
-            // The new lines revealed by the resizing are not all from the transcript. Blank the below ones.
-            for (int i = 0; i < actualShift - shiftDownOfTopRow; i++)
-                allocateFullLineIfNecessary((mScreenFirstRow + mScreenRows + i) % mTotalRows).clear(currentStyle);
-            shiftDownOfTopRow = actualShift;
-        }
-        return shiftDownOfTopRow;
-    }
-
-    private void updateOldState(int newColumns, int newRows, int newTotalRows, int[] cursor, long currentStyle) {
-        final TerminalRow[] oldLines = mLines;
-
-        final int oldActiveTranscriptRows = mActiveTranscriptRows;
-        final int oldScreenFirstRow = mScreenFirstRow;
-        final int oldScreenRows = mScreenRows;
-        final int oldTotalRows = mTotalRows;
-
-        final Cursor oldCursor = new Cursor(cursor[1], cursor[0]);
-
-        mLines = new TerminalRow[newTotalRows];
-        for (int i = 0; i < newTotalRows; i++)
-            mLines[i] = new TerminalRow(newColumns, currentStyle);
-
-        mTotalRows = newTotalRows;
-        mScreenRows = newRows;
-        mActiveTranscriptRows = mScreenFirstRow = 0;
-        mColumns = newColumns;
-
-        Cursor newCursor = new Cursor(-1, -1);
-
-        Cursor currentOutputExternal = new Cursor(0, 0);
-
-        // Loop over every character in the initial state.
-        // Blank lines should be skipped only if at end of transcript (just as is done in the "fast" resize), so we
-        // keep track how many blank lines we have skipped if we later on find a non-blank line.
-        int skippedBlankLines = 0;
-        for (int externalOldRow = -oldActiveTranscriptRows; externalOldRow < oldScreenRows; externalOldRow++) {
-            final TerminalRow oldLine = getOldTerminalRow(oldLines, oldScreenFirstRow, oldTotalRows, externalOldRow);
-            final boolean isCursorAtThisRow = externalOldRow == oldCursor.getRow();
-
-            // The cursor may only be on a non-null line, which we should not skip:
-            final boolean isOldCursorAtThisRow = (newCursor.getRow() == -1) && isCursorAtThisRow;
-            final boolean isCursorAtBlankLine = !isOldCursorAtThisRow && oldLine.isBlank();
-            if (oldLine == null || isCursorAtBlankLine) {
-                skippedBlankLines++;
-                continue;
-            }
-
-            if (skippedBlankLines > 0) {
-                // After skipping some blank lines we encounter a non-blank line. Insert the skipped blank lines.
-                insertSkippedBlankLines(currentStyle, currentOutputExternal, skippedBlankLines);
-                skippedBlankLines = 0;
-            }
-
-            int lastNonSpaceIndex = 0;
-            boolean isJustToCursor = false;
-            if (isCursorAtThisRow || oldLine.mLineWrap) {
-                // Take the whole line, either because of cursor on it, or if line wrapping.
-                lastNonSpaceIndex = oldLine.getSpaceUsed();
-                if (isCursorAtThisRow) isJustToCursor = true;
-            } else {
-                for (int i = 0; i < oldLine.getSpaceUsed(); i++)
-                    // NEWLY INTRODUCED BUG! Should not index oldLine.mStyle with char indices
-                    if (oldLine.mText[i] != ' '/* || oldLine.mStyle[i] != currentStyle */)
-                        lastNonSpaceIndex = i + 1;
-            }
-
-            copyLine(currentStyle, oldCursor, newCursor, currentOutputExternal, externalOldRow, oldLine, lastNonSpaceIndex, isJustToCursor);
-
-            // Old row has been copied. Check if we need to insert newline if old line was not wrapping:
-            final boolean isOldRowCopied = externalOldRow != (oldScreenRows - 1);
-            final boolean isOldLineNotWrapping = !oldLine.mLineWrap;
-            if (isOldRowCopied && isOldLineNotWrapping) {
-                setNewCursorRow(currentStyle, newCursor, currentOutputExternal);
-            }
-        }
-
-        cursor[0] = newCursor.getColumn();
-        cursor[1] = newCursor.getRow();
-    }
-
-    private void copyLine(long currentStyle, Cursor oldCursor, Cursor newCursor, Cursor currentOutputExternal, int externalOldRow, TerminalRow oldLine, int lastNonSpaceIndex, boolean isJustToCursor) {
-        int currentOldCol = 0;
-        long styleAtCol = 0;
-        for (int i = 0; i < lastNonSpaceIndex; i++) {
-            // Note that looping over java character, not cells.
-            final char c = oldLine.mText[i];
-            final int codePoint = (Character.isHighSurrogate(c)) ? Character.toCodePoint(c, oldLine.mText[++i]) : c;
-            final int displayWidth = WcWidth.width(codePoint);
-            // Use the last style if this is a zero-width character:
-            if (displayWidth > 0) styleAtCol = oldLine.getStyle(currentOldCol);
-
-            // Line wrap as necessary:
-            lineWrap(currentStyle, newCursor, currentOutputExternal, displayWidth);
-
-            final int offsetDueToCombiningChar = ((displayWidth <= 0 && currentOutputExternal.getColumn() > 0) ? 1 : 0);
-            final int outputColumn = currentOutputExternal.getColumn() - offsetDueToCombiningChar;
-            setChar(outputColumn, currentOutputExternal.getRow(), codePoint, styleAtCol);
-
-            if (displayWidth > 0) {
-                if (oldCursor.getRow() == externalOldRow && oldCursor.getColumn() == currentOldCol) {
-                    newCursor.setCursor(currentOutputExternal);
-                }
-                currentOldCol += displayWidth;
-                currentOutputExternal.addToColumn(displayWidth);
-                if (isJustToCursor && newCursor.getRow() != -1) break;
-            }
-        }
-    }
-
-    private void lineWrap(long currentStyle, Cursor newCursor, Cursor currentOutputExternal, int displayWidth) {
-        final boolean isLineLong = currentOutputExternal.getColumn() + displayWidth > mColumns;
-        if (isLineLong) {
-            setLineWrap(currentOutputExternal.getRow());
-            setNewCursorRow(currentStyle, newCursor, currentOutputExternal);
-        }
-    }
-
-    private TerminalRow getOldTerminalRow(TerminalRow[] oldLines, int oldScreenFirstRow, int oldTotalRows, int externalOldRow) {
-        // Do what externalToInternalRow() does but for the old state:
-        int internalOldRow = oldScreenFirstRow + externalOldRow;
-        internalOldRow = (internalOldRow < 0) ? (oldTotalRows + internalOldRow) : (internalOldRow % oldTotalRows);
-        return oldLines[internalOldRow];
-    }
-
-    private void insertSkippedBlankLines(long currentStyle, Cursor currentOutputExternal, int skippedBlankLines) {
-        for (int i = 0; i < skippedBlankLines; i++) {
-            if (currentOutputExternal.getRow() == mScreenRows - 1) {
-                scrollDownOneLine(0, mScreenRows, currentStyle);
-            } else {
-                currentOutputExternal.addToRow(1);
-            }
-            currentOutputExternal.setColumn(0);
-        }
-    }
-
-    private void setNewCursorRow(long currentStyle, Cursor newCursor, Cursor currentOutputExternal) {
-        if (currentOutputExternal.getRow() == mScreenRows - 1) {
-            if (newCursor.getRow() != -1) newCursor.addToRow(-1);
-            scrollDownOneLine(0, mScreenRows, currentStyle);
-        } else {
-            currentOutputExternal.addToRow(1);
-        }
-        currentOutputExternal.setColumn(0);
     }
 
     /**
